+++ movieland_web/src/main/java/com/solovey/movieland/web/controller/MovieController.java
@@ -8,8 +8,8 @@ import com.solovey.movieland.service.MovieService;
 import com.solovey.movieland.web.util.dto.ToDtoConverter;
 import com.solovey.movieland.web.util.currency.CurrencyService;
 import com.solovey.movieland.web.util.json.JsonJacksonConverter;
-import com.solovey.movieland.web.util.security.Protected;
-import com.solovey.movieland.web.util.security.entity.PrincipalUser;
+import com.solovey.movieland.web.security.Protected;
+import com.solovey.movieland.web.security.entity.PrincipalUser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
