+++ movieland_web/src/main/java/com/solovey/movieland/web/controller/ReviewController.java
@@ -4,8 +4,8 @@ import com.solovey.movieland.dao.enums.UserRole;
 import com.solovey.movieland.entity.Review;
 import com.solovey.movieland.service.ReviewService;
 import com.solovey.movieland.web.util.json.JsonJacksonConverter;
-import com.solovey.movieland.web.util.security.Protected;
-import com.solovey.movieland.web.util.security.entity.PrincipalUser;
+import com.solovey.movieland.web.security.Protected;
+import com.solovey.movieland.web.security.entity.PrincipalUser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
