+++ movieland_web/src/main/java/com/solovey/movieland/web/security/SecurityFilter.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security;
+package com.solovey.movieland.web.security;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,7 +7,6 @@ import javax.servlet.*;
 import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @WebFilter("/v1/*")
