+++ movieland_web/src/main/java/com/solovey/movieland/web/security/SecurityHttpRequestWrapper.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security;
+package com.solovey.movieland.web.security;
 
 
 import javax.servlet.http.HttpServletRequest;
