+++ movieland_web/src/main/java/com/solovey/movieland/web/security/entity/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security.entity;
+package com.solovey.movieland.web.security.entity;
 
 public class LoginRequest {
     private String email;
