+++ movieland_web/src/main/java/com/solovey/movieland/web/security/entity/PrincipalUser.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security.entity;
+package com.solovey.movieland.web.security.entity;
 
 import java.security.Principal;
