+++ movieland_web/src/main/java/com/solovey/movieland/web/security/entity/UserToken.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security.entity;
+package com.solovey.movieland.web.security.entity;
 
 
 public class UserToken {
