+++ movieland_web/src/main/java/com/solovey/movieland/web/security/exceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security.exceptions;
+package com.solovey.movieland.web.security.exceptions;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException() {
