+++ movieland_web/src/main/java/com/solovey/movieland/web/security/exceptions/UserTokenExpiredException.java
@@ -1,4 +1,4 @@
-package com.solovey.movieland.web.util.security.exceptions;
+package com.solovey.movieland.web.security.exceptions;
 
 public class UserTokenExpiredException extends RuntimeException {
     public UserTokenExpiredException(){
