+++ movieland_web/src/test/java/com/solovey/movieland/web/controller/ReviewControllerTest.java
@@ -7,9 +7,9 @@ import com.solovey.movieland.entity.User;
 import com.solovey.movieland.service.ReviewService;
 import com.solovey.movieland.service.UserService;
 import com.solovey.movieland.web.util.json.JsonJacksonConverter;
-import com.solovey.movieland.web.util.security.AuthenticationService;
-import com.solovey.movieland.web.util.security.entity.PrincipalUser;
-import com.solovey.movieland.web.util.security.entity.UserToken;
+import com.solovey.movieland.web.security.AuthenticationService;
+import com.solovey.movieland.web.security.entity.PrincipalUser;
+import com.solovey.movieland.web.security.entity.UserToken;
 import org.junit.Test;
 import org.springframework.test.web.servlet.MockMvc;
