+++ movieland_web/src/test/java/com/solovey/movieland/web/security/AuthenticationServiceTest.java
@@ -1,12 +1,12 @@
-package com.solovey.movieland.web.util.auth;
+package com.solovey.movieland.web.security;
 
 
 import com.solovey.movieland.entity.User;
 import com.solovey.movieland.service.UserService;
-import com.solovey.movieland.web.util.security.AuthenticationService;
-import com.solovey.movieland.web.util.security.cache.UserTokenCache;
-import com.solovey.movieland.web.util.security.entity.LoginRequest;
-import com.solovey.movieland.web.util.security.entity.UserToken;
+import com.solovey.movieland.web.security.AuthenticationService;
+import com.solovey.movieland.web.security.cache.UserTokenCache;
+import com.solovey.movieland.web.security.entity.LoginRequest;
+import com.solovey.movieland.web.security.entity.UserToken;
 import com.solovey.movieland.web.util.json.JsonJacksonConverter;
 import org.junit.Test;
